ECPINT-1468-SFRA-webpack-not-working-BUG
- moved all the js and css files from static to client folder
- client JS files to work with webpack
- updated package.json with dependencies necessary for webpack, modified its paths for correct functioning of webpack
