+++ Cartridges/int_checkoutcom_sfra/cartridge/client/default/js/apm/ideal.js
@@ -1,0 +1,17 @@
+'use strict';
+
+module.exports = {
+    formValidation: function() {
+        // Errors count
+        var errors = [];
+    
+        // Field 1 validation
+        var field1 = '#ideal_bic';
+        if ($(field1).val() === '') {
+            errors.push(field1);
+        }
+    
+        return errors;
+    }
+}
+
