+++ Cartridges/int_checkoutcom_sfra/cartridge/client/default/js/apm/qpay.js
@@ -1,0 +1,16 @@
+'use strict';
+
+module.exports = {
+    formValidation: function() {
+        // Errors count
+        var errors = [];
+    
+        // Field 1 validation
+        var field1 = '#qpay_national_id';
+        if ($(field1).val() === '') {
+            errors.push(field1);
+        }
+    
+        return errors;
+    }
+}
