+++ Cartridges/int_checkoutcom_sfra/cartridge/client/default/js/applepay.js
@@ -8,25 +8,23 @@ document.addEventListener('DOMContentLoaded', function() {
     launchApplePay();
 }, false);
 
-function initCheckoutcomApplePayValidation() {
-    $('button.submit-payment').off('click touch').one('click touch', function(e) {
-        if ($('input[name="dwfrm_billing_paymentMethod"]').val() === 'CHECKOUTCOM_APPLE_PAY') {
-            // Prevent the default button click behaviour
-            e.preventDefault();
-            e.stopImmediatePropagation();
-
-            // Validate the payment data
-            var field1 = $('input[name="dwfrm_billing_applePayForm_ckoApplePayData"]');
-            if (field1.val() === '') {
-                $('#apple-pay-content .invalid-field-message').text(
-                    window.ckoLang.applePayDataInvalid
-                );
-            } else {
-                $(this).trigger('click');
-            }
+$('button.submit-payment').off('click touch').on('click touch', function(e) {
+    if ($('input[name="dwfrm_billing_paymentMethod"]').val() === 'CHECKOUTCOM_APPLE_PAY') {
+        // Prevent the default button click behaviour
+        e.preventDefault();
+        e.stopImmediatePropagation();
+        
+        // Validate the payment data
+        var field1 = $('input[name="dwfrm_billing_applePayForm_ckoApplePayData"]');
+        if (field1.val() === '') {
+            $('#apple-pay-content .invalid-field-message').text(
+                window.ckoLang.applePayDataInvalid
+            );
+        } else {
+            $(this).trigger('click');
         }
-    });
-}
+    }
+});
 
 function getLineItems() {
     return [];
