+++ Cartridges/int_checkoutcom_sfra/cartridge/client/default/js/cardPayment.js
@@ -1,5 +1,7 @@
 'use strict';
 
+require('./checkoutcom.js');
+
 function initCheckoutcomCardValidation() {
     // Is card payment
     var condition1 = $('input[name="dwfrm_billing_paymentMethod"]').val() === 'CHECKOUTCOM_CARD';
@@ -18,7 +20,7 @@ function initCheckoutcomCardValidation() {
 function cardFormValidation() {
     $('button.submit-payment').off('click touch').on('click touch', function(e) {
     // Reset the form error messages
-        resetFormErrors();
+    checkoutcom.resetFormErrors();
 
         // Prepare the errors array
         var cardFields = [];
@@ -37,7 +39,7 @@ function cardFormValidation() {
 
         // Handle errors
         $.each(cardFields, function(i, field) {
-            if (field && field.error === 1) {
+            if (field.error === 1) {
                 $('#' + field.id).next('.invalid-feedback').show();
             }
         });
@@ -81,7 +83,7 @@ function checkCardExpirationMonth() {
     };
 
     // Check expiration month
-    if (targetField.val() === '' || ($('#expirationYear').val() == new Date().getFullYear() && targetField.val() < new Date().getMonth() + 1)) {
+    if (targetField.val() === '') {
         $('.dwfrm_billing_creditCardFields_expirationMonth .invalid-field-message').text(
             window.ckoLang.cardExpirationMonthInvalid
         );
@@ -141,12 +143,12 @@ function getFormattedNumber(num) {
  */
 function savedCardFormValidation() {
     // Enable the saved card selection
-    savedCardSelection();
+    require('./savedCardPayment').savedCardSelection();
 
     // Submit event
     $('button.submit-payment').off('click touch').one('click touch', function(e) {
     // Reset the form error messages
-        resetFormErrors();
+        checkoutcom.resetFormErrors();
 
         // Prepare some variables
         var savedCard = $('.saved-payment-instrument');
@@ -179,3 +181,9 @@ function savedCardFormValidation() {
         });
     });
 }
+
+
+document.addEventListener('DOMContentLoaded', function() {
+    initCheckoutcomCardValidation();
+});
+
