+++ Cartridges/int_checkoutcom_sfra/cartridge/client/default/js/checkout.js
@@ -1,6 +1,8 @@
 'use strict';
 
 require('base/checkout');
+var processInclude = require('base/util');
+processInclude(require('base/components/toolTip'));
 
 $(document).ready(function () {
     if (window.dw &&
