+++ Cartridges/int_checkoutcom_sfra/cartridge/client/default/scss/apm.scss
