--- Cartridges/int_checkoutcom_sfra/cartridge/static/default/js/apm/ideal.js
@@ -1,14 +1,0 @@
-'use strict';
-
-function idealFormValidation() {
-    // Errors count
-    var errors = [];
-
-    // Field 1 validation
-    var field1 = '#ideal_bic';
-    if ($(field1).val() === '') {
-        errors.push(field1);
-    }
-
-    return errors;
-}
