--- Cartridges/int_checkoutcom_sfra/cartridge/static/default/js/apm/qpay.js
@@ -1,14 +1,0 @@
-'use strict';
-
-function qpayFormValidation() {
-    // Errors count
-    var errors = [];
-
-    // Field 1 validation
-    var field1 = '#qpay_national_id';
-    if ($(field1).val() === '') {
-        errors.push(field1);
-    }
-
-    return errors;
-}
