+++ Cartridges/int_checkoutcom_sfra/cartridge/templates/default/checkout/billing/paymentOptions/forms/apm/boleto.isml
@@ -1,7 +1,6 @@
 <!-- Boleto APM option -->
 <isscript>
     var assets = require('*/cartridge/scripts/assets.js');
-    assets.addJs('/js/apm/boleto.js');
     assets.addCss('/css/apm/boleto.css');
 </isscript>
